Factor window class from split area parsing and storage
Now the work to subdivide the window is done separately in the Lua
implementation side.

The window class only keep a list of drawing areas without any hierarchy
information.

The API of attach method changed as it now only accept an integer, the
drawing are id instead of a string.